+++ CHANGELOG.md
@@ -1,2 +1,1 @@
-* Fixed LibDBIcon-1.0 icons from minimap.
-* Update to 9.0.2
+* Updated grab buttons system.